Removed unused import
