+++ resources/lint/profiles/mercury.raml
@@ -55,6 +55,6 @@ validations:
   version-format: mercury-standards.version-format
 
   #datatype rules
-  # upper-camelcase-datatype: mercury-standards.upper-camelcase-datatype
+  upper-camelcase-datatype: mercury-standards.upper-camelcase-datatype
