+++ resources/lint/profiles/slas.raml
@@ -54,4 +54,5 @@ validations:
   unique-display-names: mercury-standards.unique-display-names
   version-format: mercury-standards.version-format
   snakecase-query-parameters: slas-standards.snakecase-query-parameters
+  upper-camelcase-datatype: mercury-standards.upper-camelcase-datatype
